+++ slidez-core/build.gradle
@@ -27,7 +27,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	//RD stuff
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	//implementation 'org.flywaydb:flyway-core'
+	implementation 'org.flywaydb:flyway-core'
 	implementation 'javax.validation:validation-api:2.0.1.Final'
 	implementation 'org.passay:passay:1.0'
 	compileOnly "io.jsonwebtoken:jjwt-api:${jjwt}"
@@ -35,7 +35,6 @@ dependencies {
 			"io.jsonwebtoken:jjwt-jackson:${jjwt}"
 
 	runtimeOnly 'org.postgresql:postgresql'
-	runtimeOnly 'com.h2database:h2:1.4.200'
 
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
