--- slidez-core/src/main/java/com/binarystudio/academy/slidez/domain/user/User.java
@@ -1,27 +1,0 @@
-package com.binarystudio.academy.slidez.domain.user;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.util.UUID;
-
-@Data
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private UUID id;
-
-    private String email;
-    private String nickname;
-    private String passwordHash;
-}
