+++ slidez-core/src/main/java/com/binarystudio/academy/slidez/domain/user/UserRepository.java
@@ -1,5 +1,6 @@
 package com.binarystudio.academy.slidez.domain.user;
 
+import com.binarystudio.academy.slidez.domain.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,5 +11,5 @@ import java.util.UUID;
 public interface UserRepository extends JpaRepository<User, UUID> {
     boolean existsByEmail(String email);
 
-    Optional<User> getByNickname(String nickname);
+    Optional<User> findByEmail(String email);
 }
