+++ slidez-core/src/main/java/com/binarystudio/academy/slidez/infrastructure/security/auth/AuthController.java
@@ -1,12 +1,14 @@
 package com.binarystudio.academy.slidez.infrastructure.security.auth;
 
-import com.binarystudio.academy.slidez.domain.user.UserDto;
+import com.binarystudio.academy.slidez.domain.user.dto.UserDto;
 import com.binarystudio.academy.slidez.infrastructure.security.auth.model.AuthResponse;
 import com.binarystudio.academy.slidez.infrastructure.security.auth.model.AuthorizationRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -20,11 +22,13 @@ public class AuthController {
 	}
 
 	@PostMapping("login")
+	@ResponseStatus(HttpStatus.OK)
 	public AuthResponse login(@RequestBody AuthorizationRequest authorizationRequest) {
 		return authService.performLogin(authorizationRequest);
 	}
 
 	@PostMapping("register")
+	@ResponseStatus(HttpStatus.OK)
 	public AuthResponse register(@RequestBody UserDto userDto) {
 		return authService.register(userDto);
 	}
