+++ slidez-core/src/main/java/com/binarystudio/academy/slidez/infrastructure/security/auth/model/AuthorizationRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class AuthorizationRequest {
-	private String username;
+	private String email;
 	@ValidPassword
 	private String password;
 }
