--- slidez-core/src/main/java/com/binarystudio/academy/slidez/user/UserRepository.java
@@ -1,11 +1,0 @@
-package com.binarystudio.academy.slidez.user;
-
-import com.binarystudio.academy.slidez.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, UUID> {
-}
