--- slidez-core/src/main/java/com/binarystudio/academy/slidez/user/UserService.java
@@ -1,12 +1,0 @@
-package com.binarystudio.academy.slidez.user;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-}
