--- slidez-core/src/main/java/com/binarystudio/academy/slidez/user/dto/UserDto.java
@@ -1,18 +1,0 @@
-package com.binarystudio.academy.slidez.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDto {
-    private UUID id;
-    private String email;
-    private String firstName;
-    private String lastName;
-    private String password;
-}
