+++ slidez-core/src/main/resources/application.properties
@@ -14,3 +14,9 @@ spring.jpa.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.show-sql=true
 spring.flyway.enabled=true
+
+jwt.access.Secret=x-auth-token
+jwt.refresh.Secret=x-refresh-token
+jwt.secret = bb0wZaRRZe8iunThv/SBtIB6lNLnshEYOJzn5ky80aApUNaQhF5/ns7sEF9GfEq4emCjUd3xFq0oPdhVvB2/VA==
+jwt.secs_to_expire_access = 86400
+jwt.secs_to_expire_refresh = 2592000
